Replace yyjson with nlohmann/json
This is _not_ optimal _at all_. nlohmann/json is fantastic, but it isn't
designed for anywhere near this much data.
Unfortunately, yyjson seems to have trouble with the output from the XML
parser, while nlohmann/json doesn't.
Even with UTF-8 validation disabled, yyjson died on non-breaking spaces
and emojis, and I do not understand why.
When preparing a demo for a bug report, I was unable to reproduce with
the _exact same input strings_.
I imagine there's something weird with the interaction between the two,
but I'm unable to determine what, so I can't open a bug report.

I don't blame yyjson, for the record; this was not a problem with the
previous data dump, so something new in SE's 2024-06-30 data dump is
completely fucked, and only nlohmann/json seems able to deal with it.
No amount of string sanitising seems to be able to deal with it, and I
don't understand why.

I'm not sure if it really matters anyway. Yes, significantly worse
throughput from nlohmann/json, but the vast majority of the time
consumed does not go to JSON generation; it goes to zipping up tens of
gigabytes of data.