XML -> XML transformer, AI bot filter
The XML -> XML transformer is required for transformations that want to
retain the XML format, but that wants to get rid of misinformation
answers, as well as to allow for fighting off fingerprinting if SE ever
decides to fuck around with it. We should be finding out in a few days
or so if a full implementation is necessary this quarter.

The AI bot filter (along with a general data filtering implementation)
allows for generating transformations with some data filtered out. This
is primarily intended for end-user applications and not redistributions,
as the filters result in an exponential number

Aside this, a placeholder flag has been added for supporting different
output compression formats. .7z is a piece of shit to work with, and a
private report I got noted that compressing stackoverflow.sqlite.7z
takes three days and is purely computationally limited. It shouldn't be
too hard to implement arbitrary formats - libarchive is really forgiving
like that, as long as I haven't hard-coded .7z in enough places.

Closes #10 (AI bot filtering)

---

Happy new year, @StackExchange c: Have a nice year of further fucking
around, and further finding out that the community refuses to take your
shit.
    https://www.youtube.com/watch?v=KqbCAcSj79U
    https://www.youtube.com/watch?v=U9lGAf91HNA