Add incomplete, non-functional anti-bot-detection countermeasures
Stack Overflow has decided to block some entire countries from accessing
SO with automation tools, which affects the data dump downloader. This
means that in at least Singapore (and likely more countries, this is
just the only country where this has been verified so far), clicking
login in selenium instantly slaps the request with a full-page
cloudflare block. Resolving the captcha results in a loop, even when
this is done by a human, for reasons I do not understand.

TL;DR: SE has configured cloudflare to obliterate all forms of bots from
certain countries. If this spreads world-wide, the data dump downloader
breaks entirely.

The changes in this commit are enough to trick
    https://fingerprintjs.github.io/BotD/main/
into reporting "not a bot", but it is not good enough for cloudflare.

The only potential viable workaround is building geckodriver from
source, and _fuck that_. CF is likely detecting selenium in additional
ways that I currently cannot find.

As another weird bug, this change seems to break layouts of pages.
Though this is mostly not a problem, it breaks Google captchas entirely,
because the submit button is pushed somewhere unclickable.

I'm not sure if this is even a viable start, or if there's something
else that needs to be done, but fixing this is likely going to be far
more involved than writing some python code for end-use.
undetected-chromium, for example, downloads and patches chromium to
remove some of the bot detecting features. The equivalent package for
firefox was archived 4 days ago (15.04.2025), so that won't be a simple
fix. It was also linux-only in the first place, which makes it unviable
for this use.